\name{tempfile}
\title{Create Unique Names for (Temporary) Files}
\usage{
tempfile(pattern = "file")
}
\alias{tempfile}
\arguments{
  \item{pattern}{a character vector with the beginnings of the returned
    file names.}
}
\value{
  A character vector giving names of possible (temporary) files.  The
  names are certain to be unique between subsequent calls.  Note that no
  files are generated by \code{tempfile}.
}
\details{
    The filenames will be in the directory given by the first found of
    the environment variables \code{TMP}, \code{TEMP} and \code{R_HOME}
    (see \code{\link{Rconsole}}).
}
\seealso{\code{\link{unlink}} for deleting files.}

\keyword{file}
